Extract table export action to a top level class
Extract the action from TableExportButton so that it can be reused in
other UI components (e.g. pop up menus).
